Remove non-cray packages from site package repo (#37)
Removes non-cray packages from the site repo, as the latest ones will be
available in spack 1.0 (through spack-packages). I think there's less
need to keep them here in the site repo going forward as
- with spack 1.0 we'll be able to use newer packages more easily without
using a newer version of spack itself
- many recipes already include custom versions of the packages they
need, when they need it, and I feel like that's a more localized (and
better) way to use custom package definitions

This keeps the custom nvhpc package definition in the site repo for now.
Longer term we can do one of the following:
- don't do anything and just use the three-cuda upstream version (yes,
they're big, but squashfs' are compressed and it's not the end of the
world)
- upstream a change or variant that uses the single-cuda versions
- use a custom nvhpc package definition in the prgenv-nvfortran package
recipe
- leave the custom recipe here in alps-cluster-config