Update external packages spec (#66)
With https://github.com/spack/spack/pull/51118 external specs are
treated in a slightly different way.

In particular. An external package will be mapped 1:1 with a specific
spec, which implies that an external entry will not anymore act as a
jolly in terms of unspecified variants, but it will just represent the
spec with the set of variants as they result from the configuration file
(explicitly or implicitly). Moreover, it will be possible (required?) to
specify also dependencies for the external packages.

This PR aims at anticipating potential problems that might arise with
this change.

- `xpmem` by default has `+kernel-module`, but this would require also
specifying the `gcc` dependency. since no one seems to care about this
variant, the proposed solution is to explicitly disable this variant
`~kernel-module`