Update OpenMPI and libfabric preferences (#68)
This PR:
- Makes libfabric version >= 2.2 the preferred choice, from upstream
source. I've set no upper bound since we really do want to be testing
the latest versions as they drop in. If anything we could set a
constraint not to use libfabric 3 if/when that release happens as that
is likely to need more testing.
- Removes the custom openmpi package, all changes are upstream and e.g.
in https://github.com/eth-cscs/alps-uenv/pull/263 I anyway want to use a
very recent spack-packages commit